+++ testing.js
@@ -2,11 +2,11 @@
 import {
   ANSI,
   CliRenderer
-} from "./index-KVNO7LZO.js";
+} from "./index-AJ7ASKHW.js";
 import {
   SystemClock,
   TreeSitterClient
-} from "./index-FJESUMWG.js";
+} from "./index-SYMAOMWA.js";
 
 // src/testing/mock-keys.ts
 import { Buffer as Buffer2 } from "node:buffer";